update: fixing one decimal heading links in index, take 2
For future reference, one decimal headings in index are written at H6 in markdown - and all 'H' get their own links generated. Hence is heading in index and heading in doc are exactly the same, collision occurs. This pushes the link of the heading in doc to the syntax "link of heading in index" followed by 1. Problem is that in local view. the 1 is appended with an underscore, while in github, it's appended with a hyphen (hence incompatibility)

Therefore, the working solution  is to slightly modify the one decimal headings in the index so that collision never occurs. (like add/remove 'the', or 's', etc)

[The problem doesn't happen with 2 decimal headings of the doc as they are not presented in the index as headings (but as bullet points) hence their own links don't get generated]